+++ Sources/MapboxDirections/MapMatching/Match.swift
@@ -1,8 +1,4 @@
 import Foundation
-#if canImport(CoreLocation)
-import CoreLocation
-#endif
-import Polyline
 import Turf
 
 /**
@@ -65,7 +61,7 @@ open class Match: DirectionsResult {
      - parameter confidence: A number between 0 and 1 that indicates the Map Matching API’s confidence that the match is accurate. A higher confidence means the match is more likely to be accurate.
      - parameter weight: A `Weight` enum, which represents the weight given to a specific `Match`.
      */
-    public init(legs: [RouteLeg], shape: LineString?, distance: CLLocationDistance, expectedTravelTime: TimeInterval, confidence: Float, weight: Weight) {
+    public init(legs: [RouteLeg], shape: LineString?, distance: LocationDistance, expectedTravelTime: TimeInterval, confidence: Float, weight: Weight) {
         self.confidence = confidence
         self.weight = weight
         super.init(legs: legs, shape: shape, distance: distance, expectedTravelTime: expectedTravelTime)