+++ Tests/MapboxDirectionsTests/OfflineDirectionsTests.swift
@@ -41,8 +41,8 @@ class OfflineDirectionsTests: XCTestCase {
 
     func testDownloadTiles() {
         let directions = Directions(credentials: BogusCredentials)
-        let bounds = BoundingBox(CLLocationCoordinate2D(latitude: 37.7890, longitude: -122.4337),
-                                 CLLocationCoordinate2D(latitude: 37.7881, longitude: -122.4318))
+        let bounds = BoundingBox(southWest: CLLocationCoordinate2D(latitude: 37.7890, longitude: -122.4337),
+                                 northEast: CLLocationCoordinate2D(latitude: 37.7881, longitude: -122.4318))
         
         let version = "2018-10-16"
         let downloadExpectation = self.expectation(description: "Download tile expectation")